Complete the modern problems draft
I'm procastinating releasing this for too long. I'll make some typo
changes and then I'll just release it as-is. I'm now a firm believe that
my perfectionism always holds me back and the only way to fight this is
to try and publish _incomplete_ stuff, or at least stuff I'm not sure
I'm satisfied with yet.